requirements.txt
# Core Streamlit and Database
streamlit>=1.28.0
supabase>=2.0.0
python-dotenv>=1.0.0

# Authentication and Security
bcrypt>=4.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# File Processing
Pillow>=10.0.0
openpyxl>=3.1.0
python-docx>=0.8.11

# Visualization (Optional but recommended)
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# AI Integration (Optional - for auto grading)
google-generativeai>=0.3.0

# Development Tools (Optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# HTTP Requests (nếu cần)
requests>=2.31.0

# Date/Time utilities (nếu cần)
python-dateutil>=2.8.0